requirements.txt
feedparser
keyboard
pyttsx3
pvporcupine
pyaudio
spotipy
requests
newsapi-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
SpeechRecognition
spacy
google-generativeai
